requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
flask==2.2.2
flask-cors==3.0.10
phe==1.5.0
pycryptodome==3.16.0
